major refactoring - main OS is now in the edelage folder
A lot of files created. Dump function is working OK. Init task
exists but so far doesn't do much. Scheduling works OK but some
obvious bugs remains. Very simple syscalls which more or less
only works for the console but it is used to test the printk part.
Printk works OK, at least enough to continue with other stuff.

Next steps are to create a nice device layer concept and adapt
the console driver to it.